+++ lib/google_static_maps_controller.dart
@@ -1,16 +1,13 @@
 library google_static_maps_controller;
 
-import 'dart:ui';
 import 'dart:math';
 import 'package:flutter/widgets.dart';
 
-part './src/widgets/static_map.dart';
+/// data models
 part './src/data/enums.dart';
 part './src/data/location.dart';
-part './src/data/map_part.dart';
-
-part './src/utils/utils.dart';
-part './src/controller.dart';
+part './src/data/encodable_url_part.dart';
+part './src/data/geocoded_location.dart';
 
 /// markers
 part './src/markers/marker.dart';
@@ -26,3 +23,10 @@ part './src/style/data/visibility_rule.dart';
 
 /// paths
 part './src/paths/path.dart';
+
+/// utils
+part './src/utils/polyline_encoder.dart';
+part './src/utils/utils.dart';
+
+part './src/controller.dart';
+part './src/widgets/static_map.dart';