+++ lib/src/data/encodable_url_part.dart
@@ -1,5 +1,5 @@
 part of google_static_maps_controller;
 
-abstract class MapPart {
+abstract class EncodableUrlPart {
   String toUrlString();
 }