+++ lib/src/data/location.dart
@@ -1,29 +1,98 @@
 part of google_static_maps_controller;
 
-class Location implements MapPart {
+@immutable
+class Location implements EncodableUrlPart, GeocodedLocation {
   const Location(double latitude, double longitude)
       : latitude =
             latitude < -90.0 ? -90.0 : (90.0 < latitude ? 90.0 : latitude),
         longitude = (longitude + 180.0) % 360.0 - 180.0;
 
   final double latitude;
-
   final double longitude;
 
   @override
-  String toString() => '$runtimeType($latitude, $longitude)';
+  bool operator ==(Object other) =>
+      other is Location &&
+      other.latitude == latitude &&
+      other.longitude == longitude;
+
+  Location operator -(Object object) {
+    if (object is num) {
+      return Location(
+        latitude - object,
+        longitude - object,
+      );
+    } else if (object is Location) {
+      return Location(
+        latitude - object.latitude,
+        longitude - object.longitude,
+      );
+    } else {
+      throw UnsupportedError("Unsupported operation.");
+    }
+  }
+
+  Location operator *(Object object) {
+    if (object is num) {
+      return Location(
+        latitude * object,
+        longitude * object,
+      );
+    } else if (object is Location) {
+      return Location(
+        latitude * object.latitude,
+        longitude * object.longitude,
+      );
+    } else {
+      throw UnsupportedError("Unsupported operation.");
+    }
+  }
+
+  Location operator /(Object object) {
+    if (object is num) {
+      return Location(
+        latitude / object,
+        longitude / object,
+      );
+    } else if (object is Location) {
+      return Location(
+        latitude / object.latitude,
+        longitude / object.longitude,
+      );
+    } else {
+      throw UnsupportedError("Unsupported operation.");
+    }
+  }
+
+  Location operator +(Object object) {
+    if (object is num) {
+      return Location(
+        latitude + object,
+        longitude + object,
+      );
+    } else if (object is Location) {
+      return Location(
+        latitude + object.latitude,
+        longitude + object.longitude,
+      );
+    } else {
+      throw UnsupportedError("Unsupported operation.");
+    }
+  }
+
+  /// Encode this location using
+  /// [Encoded Polyline Algorithm](https://developers.google.com/maps/documentation/utilities/polylinealgorithm).
+  String encode() =>
+      "${PolylineEncoder.encode(latitude)}${PolylineEncoder.encode(longitude)}";
 
   @override
-  bool operator ==(Object object) =>
-      object is Location &&
-      object.latitude == latitude &&
-      object.longitude == longitude;
-
   int get hashCode => hashValues(latitude, longitude);
 
   @override
   String toUrlString() {
-    // TODO: remove uneccessary space between lat nad lng, then update tests
-    return "$latitude, $longitude";
+    return "$latitude,$longitude";
   }
+
+  @override
+  String toString() => '$runtimeType($latitude, $longitude)';
 }