+++ test/google_static_maps_test.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:google_static_maps_controller/google_static_maps_controller.dart';
@@ -14,7 +12,7 @@ void main() {
         googleApiKey: _mockGoogleApiKey,
         height: 300,
         width: 200,
-        center: const Location(20, 30),
+        center: GeocodedLocation.latLng(20, 30),
         zoom: 10,
         format: MapImageFormat.gif,
         language: "PL",
@@ -29,7 +27,7 @@ void main() {
         "https://maps.googleapis.com/maps/api/staticmap?"
         "key=GOOGLE_API_KEY"
         "&size=200x300"
-        "&center=20.0%2C+30.0"
+        "&center=20.0%2C30.0"
         "&language=PL"
         "&maptype=satellite"
         "&zoom=10"
@@ -42,10 +40,14 @@ void main() {
   });
 
   test("Builds link with markers correctly", () {
-    final controller = StaticMapController(
+    const controller = StaticMapController(
       googleApiKey: _mockGoogleApiKey,
       height: 300,
       width: 200,
+      visible: [
+        GeocodedLocation.address('Example address somewhere'),
+        GeocodedLocation.address('Example address 12')
+      ],
       markers: <Marker>[
         Marker(
           color: Color(0xFFFF0000),
@@ -77,7 +79,9 @@ void main() {
         // marker styles
         "&markers=size%3Amid%7Clabel%3AA%7Ccolor%3A0xff0000"
         // marker locations
-        "%7C10.0%2C+20.0%7C20.0%2C+30.0%7C20.0%2C+30.0",
+        "%7C10.0%2C20.0%7C20.0%2C30.0%7C20.0%2C30.0"
+        "&visible=Example+address+somewhere"
+        "&visible=Example+address+12",
       ),
     );
   });
@@ -92,15 +96,16 @@ void main() {
       [3.99, MapScale.scale4],
       [4.01, MapScale.scale4],
       [9.01, MapScale.scale4],
-    ])
+    ]) {
       expect(
         getScaleForDevicePixelRatio(value[0] as double),
         equals(value[1]),
       );
+    }
   });
 
   test("Encodes custom marker correctly", () {
-    final marker = Marker.custom(
+    const marker = Marker.custom(
       anchor: MarkerAnchor.top,
       icon: "www.example.com/image",
       locations: [
@@ -114,12 +119,12 @@ void main() {
       equals(
         "anchor:top"
         "|icon:www.example.com/image"
-        "|20.0, 10.0"
-        "|20.0, 20.0",
+        "|20.0,10.0"
+        "|20.0,20.0",
       ),
     );
 
-    final marker1 = Marker.custom(
+    const marker1 = Marker.custom(
       anchor: MarkerAnchor(64, 0),
       icon: "www.example.com/image",
       locations: [
@@ -132,13 +137,13 @@ void main() {
       equals(
         "anchor:64, 0"
         "|icon:www.example.com/image"
-        "|20.0, 10.0",
+        "|20.0,10.0",
       ),
     );
   });
 
   test("Build link with custom markers correctly", () {
-    final controller = StaticMapController(
+    const controller = StaticMapController(
       googleApiKey: _mockGoogleApiKey,
       height: 200,
       width: 200,
@@ -158,8 +163,8 @@ void main() {
         "&size=200x200"
         "&markers="
         "icon%3Awww.example.com%2Fimage"
-        "%7C20.0%2C+10.0"
-        "%7C20.0%2C+20.0",
+        "%7C20.0%2C10.0"
+        "%7C20.0%2C20.0",
       ),
     );
   });
@@ -172,14 +177,14 @@ void main() {
         googleApiKey: _mockGoogleApiKey,
         height: 200,
         width: 200,
-        center: Location(-3.1467579, -59.8753814),
+        center: const GeocodedLocation.latLng(-3.1467579, -59.8753814),
         zoom: 10,
         mapId: 'mockedID',
         styles: [
           MapStyle(
             feature: StyleFeature.poi.government,
             element: StyleElement.geometry.fill,
-            rules: [
+            rules: const [
               StyleRule.visibility(VisibilityRule.simplified),
               StyleRule.color(Colors.green),
             ],