+++ test/marker_test.dart
@@ -2,12 +2,10 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:google_static_maps_controller/google_static_maps_controller.dart';
 
 void main() {
-  TestWidgetsFlutterBinding.ensureInitialized();
-
   test(
       "Throws error when marker toUrlString method is called for "
       "marker with no locations provided.", () {
-    Marker marker = Marker(
+    var marker = const Marker(
       locations: <Location>[],
     );
 
@@ -16,7 +14,7 @@ void main() {
       throwsA(isA<StateError>()),
     );
 
-    marker = Marker.custom(
+    marker = const Marker.custom(
       icon: "http://example.com/icon.png",
       locations: <Location>[],
     );